ev_backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
ConfigArgParse==1.7
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
executing==2.2.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
geographiclib==2.0
geopy==2.4.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.2.1
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.3
locust==2.37.0
locust-cloud==1.21.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
polyline==2.0.2
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
pyzmq==26.4.0
razorpay==1.4.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==79.0.1
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
gunicorn